+++ frontend/src/components/navbar/NavbarLogo.js
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import Logo from '../assets/Logo.png';
+import Logo from '../../assets/Logo.png';
 
 export const NavbarLogo = () => {
     const navigate = useNavigate();
